game sim project
A semester long project that was made in the capstone class game sim.
In this project a team of students, 12 in this case, make a game from start to finish. the end goal is to present it and have a working, functional fun game.
I was working with another programmer in which we split the project about 50/50. My systems claimed were the interaction systems, inventory, shaders, some AI and  Competition systems.